Reworking to pull shared methods from API and Helcim.js into mixins Creating view to handle Helcim.js responses
